like

+++ Modules/Like/App/Http/Controllers/LikeController.php
@@ -1,7 +1,5 @@
 <?php
-
 namespace Modules\Like\App\Http\Controllers;
-
 use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -26,40 +24,45 @@ class LikeController extends Controller
         });
         return view('public.dashboard.like',compact('likeItem','menuItems'));
     }
-    public function addlike(Product $product,Request $request)
-    {
-        $user = auth()->user();
-        $productId = $request->input('product_id');
-        $product = Product::find($productId);
-    
-        if (!$product) {
-            return redirect()->back()->with('error', 'Sản phẩm không tồn tại.');
-        }
-    
-        $existingFavorite = Wishlists::where('user_id', $user->id)
-            ->where('product_id', $productId)
-            ->first();
-    
-        if ($existingFavorite) {
-            // Nếu sản phẩm đã tồn tại trong danh sách yêu thích, xóa nó
-            $existingFavorite->delete();
-            return redirect()->route('like')->with('success', 'Đã xóa sản phẩm khỏi danh sách yêu thích.');
-        } else {
-            // Nếu sản phẩm chưa tồn tại trong danh sách yêu thích, thêm nó
-            Wishlists::create([
-                'user_id' => $user->id,
-                'product_id' => $productId,
-            ]);
-            return back()->with('success', 'Đã thêm sản phẩm vào danh sách yêu thích.');
-        }
+    public function addlike(Product $product, Request $request)
+{
+    // Check if the user is authenticated
+    if (!auth()->check()) {
+        return redirect()->back()->with('error', 'Bạn cần đăng nhập để thích sản phẩm.');
     }
+
+    $user = auth()->user();
+    $productId = $request->input('product_id');
+    $product = Product::find($productId);
+
+    if (!$product) {
+        return redirect()->back()->with('error', 'Sản phẩm không tồn tại.');
+    }
+
+    $existingFavorite = Wishlists::where('user_id', $user->id)
+        ->where('product_id', $productId)
+        ->first();
+
+    if ($existingFavorite) {
+        // Nếu sản phẩm đã tồn tại trong danh sách yêu thích, xóa nó
+        $existingFavorite->delete();
+        return redirect()->route('like')->with('success', 'Đã xóa sản phẩm khỏi danh sách yêu thích.');
+    } else {
+        // Nếu sản phẩm chưa tồn tại trong danh sách yêu thích, thêm nó
+        Wishlists::create([
+            'user_id' => $user->id,
+            'product_id' => $productId,
+        ]);
+        return back()->with('success', 'Đã thêm sản phẩm vào danh sách yêu thích.');
+    }
+}
 
 public function deletelike(string $id)
 {
     $likeItem = wishlists::find($id);
     if ($likeItem) {
         $likeItem->delete();
-        return back()->with('success_message','xóa sản phẩm yêu thích thành công');
+        return back()->with('success','xóa sản phẩm yêu thích thành công');
     }else
     {
         return redirect()->route('like');
